Pull out E2E tests, add one for X01
